+++ src/main/java/com/paneedah/weaponlib/ItemStorage.java
@@ -1,6 +1,7 @@
 package com.paneedah.weaponlib;
 
 import com.paneedah.mwc.utils.ModReference;
+import com.paneedah.weaponlib.animation.Transform;
 import com.paneedah.weaponlib.crafting.CraftingEntry;
 import com.paneedah.weaponlib.crafting.CraftingGroup;
 import com.paneedah.weaponlib.crafting.CraftingRegistry;
@@ -48,7 +49,6 @@ public class ItemStorage extends Item implements ModelSource, IModernCrafting {
         private Consumer<RenderContext<RenderableState>> firstPersonLeftHandPositioning;
         private Consumer<RenderContext<RenderableState>> firstPersonRightHandPositioning;
         private int size;
-        private String guiTextureName;
         private int guiTextureWidth = DEFAULT_GUI_TEXTURE_WIDTH;
         
         private String modelFileString;
@@ -66,7 +66,52 @@ public class ItemStorage extends Item implements ModelSource, IModernCrafting {
             this.size = size;
             return this;
         }
-        
+
+        /**
+         * Sets a Predicate to filter items that can be stored in the inventory.
+         * <p>
+         * This Predicate can be tailored based on item type or specific items, enabling fine-grained control over what can be carried.
+         *
+         * <p>
+         * <b>Type-based Filtering:</b> Can be used to filter items based on their types.
+         *
+         * <ul>
+         * <li> Allow a specific type: {@code item -> item instanceof SpecificType}
+         * <li> Exclude a specific type: {@code item -> !(item instanceof SpecificType)}
+         * <li> Allow multiple types: {@code item -> item instanceof SpecificType1 || item instanceof SpecificType2}
+         * <li> Exclude multiple types: {@code item -> !(item instanceof SpecificType1 || item instanceof SpecificType2)}
+         * <li> Allow one type but exclude another: {@code item -> item instanceof SpecificType1 && !(item instanceof SpecificType2)}
+         * </ul>
+         * In a game context, the 'instanceof' operator could be applied to classes like ItemAmmo, Weapon, Armor, etc.
+         *
+         * <p>
+         * <b>Item-based Filtering:</b> Can be used to filter specific items, irrespective of their types.
+         *
+         * <ul>
+         * <li> Allow a specific item: {@code item -> item == specificItem}
+         * <li> Exclude a specific item: {@code item -> item != specificItem}
+         * <li> Allow multiple items: {@code item -> item == specificItem1 || item == specificItem2}
+         * <li> Exclude multiple items: {@code item -> item != specificItem1 && item != specificItem2}
+         * <li> Allow one item but exclude another: {@code item -> item == specificItem1 && item != specificItem2}
+         * </ul>
+         *
+         * <p>
+         * <b>Combined Filtering:</b> Can be used to combine the type and item-based filtering.
+         *
+         * <ul>
+         * <li> Allow a type and a specific item: {@code item -> item instanceof SpecificType && item == specificItem}
+         * <li> Allow a type but exclude an item: {@code item -> item instanceof SpecificType && item != specificItem}
+         * <li> Exclude a type and a specific item: {@code item -> !(item instanceof SpecificType) && item != specificItem}
+         * <li> Allow multiple types or items: {@code item -> (item instanceof SpecificType1 || item instanceof SpecificType2) || (item == specificItem1 || item == specificItem2)}
+         * <li> Exclude multiple types or items: {@code item -> !(item instanceof SpecificType1 || item instanceof SpecificType2) && item != specificItem1 && item != specificItem2}
+         * <li> Allow a type but exclude another type and a specific item: {@code item -> item instanceof SpecificType1 && !(item instanceof SpecificType2) && item != specificItem}
+         * </ul>
+         *
+         * This method is particularly useful when there's a need to strictly control the types or specific items in an inventory.
+         *
+         * @param validItemPredicate The Predicate function that defines the criteria for item validity
+         * @return Returns the Builder instance for method chaining
+         */
         public Builder withValidItemPredicate(Predicate<Item> validItemPredicate) {
             this.validItemPredicate = validItemPredicate;
             return this;
@@ -85,11 +130,6 @@ public class ItemStorage extends Item implements ModelSource, IModernCrafting {
         
         public Builder withModel(ModelBase model) {
             this.model = model;
-            return this;
-        }
-        
-        public Builder withGuiTextureName(String guiTextureName) {
-            this.guiTextureName = guiTextureName;
             return this;
         }
         
@@ -160,7 +200,7 @@ public class ItemStorage extends Item implements ModelSource, IModernCrafting {
                 .withFirstPersonPositioning(builder.firstPersonPositioning)
                 .withThirdPersonPositioning(builder.thirdPersonPositioning)
                 .withCustomEquippedPositioning(builder.customEquippedPositioning)
-                .withInventoryPositioning(builder.inventoryPositioning)
+                .withInventoryPositioning(builder.inventoryPositioning != null ? builder.inventoryPositioning : stack -> new Transform().withPosition(-0.15, -4.15, 0.35).withRotation(18, -50, 0).withScale(2.9, 2.9, 2.9).doGLDirect())
                 .withEntityModelPositioning(builder.entityModelPositioning)
                 .withFirstPersonModelPositioning(builder.firstPersonModelPositioning)
                 .withThirdPersonModelPositioning(builder.thirdPersonModelPositioning)
@@ -179,15 +219,8 @@ public class ItemStorage extends Item implements ModelSource, IModernCrafting {
             if(size <= 0) {
                 throw new IllegalStateException("ItemStorage size must be greater than 0");
             }
-            
-            if(guiTextureName == null) {
-                throw new IllegalStateException("ItemStorage gui texture not set");
-            }
-            
-            if(!guiTextureName.startsWith("textures/gui/")) {
-                guiTextureName = "textures/gui/" + guiTextureName;
-            }
-            ResourceLocation guiTextureLocation = new ResourceLocation(ModReference.id, addFileExtension(guiTextureName, ".png"));
+
+            ResourceLocation guiTextureLocation = new ResourceLocation(ModReference.id, "textures/gui/inventory/carryable/" + size + "slots.png");
             
             ItemStorage item = new ItemStorage(modContext, size, validItemPredicate, guiTextureLocation, this.guiTextureWidth);