+++ Changelog.md
@@ -34,6 +34,7 @@ and this project follows to [Ragnarök Versioning Convention](https://gist.githu
 ### Changed
 
 - Overhauled textures of AAC Honey Badger, Beowulf, HK 417, M16A1 and the pink camo skin
+- You can now carry sidearms in carryable inventories
 - Reworked the config system:
   - Old XML config has been removed, and all configs have been consolidated into a new JSON config. Issues with some non-functioning configs have been resolved
   - Adjusted craftingmappings.json to now only override recipes, rather than removing all recipes
@@ -60,6 +61,7 @@ and this project follows to [Ragnarök Versioning Convention](https://gist.githu
 - Updated `mcmod.info` description and credits
 - Migrated the config setting for bullets breaking glass into the new config file (ModernWarfareCubed.cfg)
 - Forge configuration `Emissive Item Rendering` is now forced to be off (This does not influence Optifine emissive item rendering)
+- Renamed `Armor` creative tab to `Equipment`
 
 ### Fixed
 
@@ -75,6 +77,7 @@ and this project follows to [Ragnarök Versioning Convention](https://gist.githu
 - Fixed crashes caused by certain shaders
 - Fixed duplication issues related to copper, gunpowder, and ammo
 - Fixed a problem where the carryable inventory tab was shifting the entire inventory to the left for no reasons
+- Fixed position of the backpacks in inventory
 
 ### Removed